requirements.txt
streamlit
Pillow
folium
streamlit-folium
opencv-python-headless
pandas
numpy
